backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
python-dateutil>=2.8.2

# AI/ML Dependencies
torch>=1.9.0
transformers>=4.15.0
sentencepiece>=0.1.96
protobuf>=3.20.0